requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django== 4.2.14
django-extensions==3.2.3
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.0
idna==3.7
lxml==5.3.0
more-itertools==10.4.0
pillow==10.4.0
premailer==3.10.0
requests==2.32.3
sqlparse==0.5.1
testfixtures==8.3.0
urllib3==2.2.2
